build moves only tests+benches pending
