+++ datafusion/datasource/src/empty.rs
@@ -15,27 +15,31 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! EmptyRelation produce_one_row=true execution plan
+//! EmptyRelation with produce_one_row=false execution plan
 
 use std::any::Any;
 use std::sync::Arc;
 
-use super::{common, DisplayAs, PlanProperties, SendableRecordBatchStream, Statistics};
-use crate::execution_plan::{Boundedness, EmissionType};
-use crate::{memory::MemoryStream, DisplayFormatType, ExecutionPlan, Partitioning};
+use crate::memory::MemoryStream;
+use datafusion_physical_plan::{
+    common, DisplayAs, PlanProperties, SendableRecordBatchStream, Statistics,
+};
+use datafusion_physical_plan::{
+    execution_plan::{Boundedness, EmissionType},
+    DisplayFormatType, ExecutionPlan, Partitioning,
+};
 
-use arrow::array::{ArrayRef, NullArray};
-use arrow::array::{RecordBatch, RecordBatchOptions};
-use arrow::datatypes::{DataType, Field, Fields, Schema, SchemaRef};
+use arrow::datatypes::SchemaRef;
+use arrow::record_batch::RecordBatch;
 use datafusion_common::{internal_err, Result};
 use datafusion_execution::TaskContext;
 use datafusion_physical_expr::EquivalenceProperties;
 
 use log::trace;
 
-/// Execution plan for empty relation with produce_one_row=true
+/// Execution plan for empty relation with produce_one_row=false
 #[derive(Debug, Clone)]
-pub struct PlaceholderRowExec {
+pub struct EmptyExec {
     /// The schema for the produced row
     schema: SchemaRef,
     /// Number of partitions
@@ -43,48 +47,28 @@ pub struct PlaceholderRowExec {
     cache: PlanProperties,
 }
 
-impl PlaceholderRowExec {
-    /// Create a new PlaceholderRowExec
+impl EmptyExec {
+    /// Create a new EmptyExec
     pub fn new(schema: SchemaRef) -> Self {
-        let partitions = 1;
-        let cache = Self::compute_properties(Arc::clone(&schema), partitions);
-        PlaceholderRowExec {
+        let cache = Self::compute_properties(Arc::clone(&schema), 1);
+        EmptyExec {
             schema,
-            partitions,
+            partitions: 1,
             cache,
         }
     }
 
-    /// Create a new PlaceholderRowExecPlaceholderRowExec with specified partition number
+    /// Create a new EmptyExec with specified partition number
     pub fn with_partitions(mut self, partitions: usize) -> Self {
         self.partitions = partitions;
-        // Update output partitioning when updating partitions:
+        // Changing partitions may invalidate output partitioning, so update it:
         let output_partitioning = Self::output_partitioning_helper(self.partitions);
         self.cache = self.cache.with_partitioning(output_partitioning);
         self
     }
 
     fn data(&self) -> Result<Vec<RecordBatch>> {
-        Ok({
-            let n_field = self.schema.fields.len();
-            vec![RecordBatch::try_new_with_options(
-                Arc::new(Schema::new(
-                    (0..n_field)
-                        .map(|i| {
-                            Field::new(format!("placeholder_{i}"), DataType::Null, true)
-                        })
-                        .collect::<Fields>(),
-                )),
-                (0..n_field)
-                    .map(|_i| {
-                        let ret: ArrayRef = Arc::new(NullArray::new(1));
-                        ret
-                    })
-                    .collect(),
-                // Even if column number is empty we can generate single row.
-                &RecordBatchOptions::new().with_row_count(Some(1)),
-            )?]
-        })
+        Ok(vec![])
     }
 
     fn output_partitioning_helper(n_partitions: usize) -> Partitioning {
@@ -102,7 +86,7 @@ impl PlaceholderRowExec {
     }
 }
 
-impl DisplayAs for PlaceholderRowExec {
+impl DisplayAs for EmptyExec {
     fn fmt_as(
         &self,
         t: DisplayFormatType,
@@ -110,15 +94,15 @@ impl DisplayAs for PlaceholderRowExec {
     ) -> std::fmt::Result {
         match t {
             DisplayFormatType::Default | DisplayFormatType::Verbose => {
-                write!(f, "PlaceholderRowExec")
+                write!(f, "EmptyExec")
             }
         }
     }
 }
 
-impl ExecutionPlan for PlaceholderRowExec {
+impl ExecutionPlan for EmptyExec {
     fn name(&self) -> &'static str {
-        "PlaceholderRowExec"
+        "EmptyExec"
     }
 
     /// Return a reference to Any that can be used for downcasting
@@ -146,11 +130,11 @@ impl ExecutionPlan for PlaceholderRowExec {
         partition: usize,
         context: Arc<TaskContext>,
     ) -> Result<SendableRecordBatchStream> {
-        trace!("Start PlaceholderRowExec::execute for partition {} of context session_id {} and task_id {:?}", partition, context.session_id(), context.task_id());
+        trace!("Start EmptyExec::execute for partition {} of context session_id {} and task_id {:?}", partition, context.session_id(), context.task_id());
 
         if partition >= self.partitions {
             return internal_err!(
-                "PlaceholderRowExec invalid partition {} (expected less than {})",
+                "EmptyExec invalid partition {} (expected less than {})",
                 partition,
                 self.partitions
             );
@@ -166,7 +150,7 @@ impl ExecutionPlan for PlaceholderRowExec {
     fn statistics(&self) -> Result<Statistics> {
         let batch = self
             .data()
-            .expect("Create single row placeholder RecordBatch should not fail");
+            .expect("Create empty RecordBatch should not fail");
         Ok(common::compute_record_batch_statistics(
             &[batch],
             &self.schema,
@@ -178,23 +162,39 @@ impl ExecutionPlan for PlaceholderRowExec {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{test, with_new_children_if_necessary};
+    use crate::test;
+    use crate::with_new_children_if_necessary;
+
+    #[tokio::test]
+    async fn empty() -> Result<()> {
+        let task_ctx = Arc::new(TaskContext::default());
+        let schema = test::aggr_test_schema();
+
+        let empty = EmptyExec::new(Arc::clone(&schema));
+        assert_eq!(empty.schema(), schema);
+
+        // We should have no results
+        let iter = empty.execute(0, task_ctx)?;
+        let batches = common::collect(iter).await?;
+        assert!(batches.is_empty());
+
+        Ok(())
+    }
 
     #[test]
     fn with_new_children() -> Result<()> {
         let schema = test::aggr_test_schema();
+        let empty = Arc::new(EmptyExec::new(Arc::clone(&schema)));
 
-        let placeholder = Arc::new(PlaceholderRowExec::new(schema));
-
-        let placeholder_2 = with_new_children_if_necessary(
-            Arc::clone(&placeholder) as Arc<dyn ExecutionPlan>,
+        let empty2 = with_new_children_if_necessary(
+            Arc::clone(&empty) as Arc<dyn ExecutionPlan>,
             vec![],
         )?;
-        assert_eq!(placeholder.schema(), placeholder_2.schema());
+        assert_eq!(empty.schema(), empty2.schema());
 
-        let too_many_kids = vec![placeholder_2];
+        let too_many_kids = vec![empty2];
         assert!(
-            with_new_children_if_necessary(placeholder, too_many_kids).is_err(),
+            with_new_children_if_necessary(empty, too_many_kids).is_err(),
             "expected error when providing list of kids"
         );
         Ok(())
@@ -204,44 +204,11 @@ mod tests {
     async fn invalid_execute() -> Result<()> {
         let task_ctx = Arc::new(TaskContext::default());
         let schema = test::aggr_test_schema();
-        let placeholder = PlaceholderRowExec::new(schema);
+        let empty = EmptyExec::new(schema);
 
-        // Ask for the wrong partition
-        assert!(placeholder.execute(1, Arc::clone(&task_ctx)).is_err());
-        assert!(placeholder.execute(20, task_ctx).is_err());
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn produce_one_row() -> Result<()> {
-        let task_ctx = Arc::new(TaskContext::default());
-        let schema = test::aggr_test_schema();
-        let placeholder = PlaceholderRowExec::new(schema);
-
-        let iter = placeholder.execute(0, task_ctx)?;
-        let batches = common::collect(iter).await?;
-
-        // Should have one item
-        assert_eq!(batches.len(), 1);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn produce_one_row_multiple_partition() -> Result<()> {
-        let task_ctx = Arc::new(TaskContext::default());
-        let schema = test::aggr_test_schema();
-        let partitions = 3;
-        let placeholder = PlaceholderRowExec::new(schema).with_partitions(partitions);
-
-        for n in 0..partitions {
-            let iter = placeholder.execute(n, Arc::clone(&task_ctx))?;
-            let batches = common::collect(iter).await?;
-
-            // Should have one item
-            assert_eq!(batches.len(), 1);
-        }
-
+        // ask for the wrong partition
+        assert!(empty.execute(1, Arc::clone(&task_ctx)).is_err());
+        assert!(empty.execute(20, task_ctx).is_err());
         Ok(())
     }
 }
