+++ datafusion/datasource/src/values.rs
@@ -20,18 +20,19 @@
 use std::any::Any;
 use std::sync::Arc;
 
-use super::{common, DisplayAs, PlanProperties, SendableRecordBatchStream, Statistics};
-use crate::execution_plan::{Boundedness, EmissionType};
-use crate::{
-    memory::MemoryStream, ColumnarValue, DisplayFormatType, ExecutionPlan, Partitioning,
-    PhysicalExpr,
-};
-
+use crate::memory::MemoryStream;
 use arrow::datatypes::{Schema, SchemaRef};
 use arrow::record_batch::{RecordBatch, RecordBatchOptions};
 use datafusion_common::{internal_err, plan_err, Result, ScalarValue};
 use datafusion_execution::TaskContext;
 use datafusion_physical_expr::EquivalenceProperties;
+use datafusion_physical_plan::execution_plan::{Boundedness, EmissionType};
+use datafusion_physical_plan::{
+    common, DisplayAs, PlanProperties, SendableRecordBatchStream, Statistics,
+};
+use datafusion_physical_plan::{
+    ColumnarValue, DisplayFormatType, ExecutionPlan, Partitioning, PhysicalExpr,
+};
 
 /// Execution plan for values list based relation (produces constant rows)
 #[deprecated(
@@ -230,8 +231,8 @@ impl ExecutionPlan for ValuesExec {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::expressions::lit;
-    use crate::test::{self, make_partition};
+    use datafusion_physical_plan::expressions::lit;
+    use datafusion_physical_plan::test::{self, make_partition};
 
     use arrow::datatypes::{DataType, Field};
     use datafusion_common::stats::{ColumnStatistics, Precision};
