+++ datafusion/functions-table/Cargo.toml
@@ -42,6 +42,7 @@ arrow = { workspace = true }
 async-trait = { workspace = true }
 datafusion-catalog = { workspace = true }
 datafusion-common = { workspace = true }
+datafusion-datasource = { workspace = true }
 datafusion-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
 parking_lot = { workspace = true }
