+++ datafusion/functions-table/src/generate_series.rs
@@ -23,8 +23,8 @@ use datafusion_catalog::Session;
 use datafusion_catalog::TableFunctionImpl;
 use datafusion_catalog::TableProvider;
 use datafusion_common::{plan_err, Result, ScalarValue};
+use datafusion_datasource::memory::{LazyBatchGenerator, LazyMemoryExec};
 use datafusion_expr::{Expr, TableType};
-use datafusion_physical_plan::memory::{LazyBatchGenerator, LazyMemoryExec};
 use datafusion_physical_plan::ExecutionPlan;
 use parking_lot::RwLock;
 use std::fmt;
