+++ datafusion/physical-optimizer/src/aggregate_statistics.rs
@@ -20,8 +20,8 @@ use datafusion_common::config::ConfigOptions;
 use datafusion_common::scalar::ScalarValue;
 use datafusion_common::tree_node::{Transformed, TransformedResult, TreeNode};
 use datafusion_common::Result;
+use datafusion_datasource::placeholder_row::PlaceholderRowExec;
 use datafusion_physical_plan::aggregates::AggregateExec;
-use datafusion_physical_plan::placeholder_row::PlaceholderRowExec;
 use datafusion_physical_plan::projection::ProjectionExec;
 use datafusion_physical_plan::udaf::{AggregateFunctionExpr, StatisticsArgs};
 use datafusion_physical_plan::{expressions, ExecutionPlan};
