+++ datafusion/substrait/src/physical_plan/producer.rs
@@ -24,8 +24,8 @@ use crate::variation_const::{
 
 use datafusion::arrow::datatypes::DataType;
 use datafusion::error::{DataFusionError, Result};
-use datafusion::physical_plan::source::DataSourceExec;
 use datafusion::physical_plan::{displayable, ExecutionPlan};
+use datafusion_datasource::source::DataSourceExec;
 
 use datafusion::datasource::physical_plan::{FileScanConfig, ParquetSource};
 use substrait::proto::expression::mask_expression::{StructItem, StructSelect};
